Basic attribute loading
This is "loading" and not "parsing" because the union members in
attribute_info.info are not parsed out individually. It is probably
necessary to use pack(1) on that union if we wish to avoid dispatching
the unpacking ; unlike the constant_pool dispatching, this dispatch
would be by string, which is a bit more annoying. If we do dispatch by
name, we'll probably want to use gperf ; if we don't, we might want to
backport whatever pack-solution we use here to the constant_pool parsing
as well.